requirements.txt
fastapi
uvicorn[standard]
aiohttp
openai
google-generativeai
python-dotenv

